requirements.txt
streamlit==0.79.0
torch==1.8.1
torchvision==0.9.1
Pillow==8.1.2
ultralytics
opencv-python-headless
gdown==3.13.0
requests==2.25.1
